<browser_ext/generic_build: Provide generic build>
Generic build allow to create a base ZIP which will be extended to
create extension for different browser (both chromium and firefox
extension file format is based on ZIP).

* For firefox the generated XPI is only a symlink to ZIP file.
* For Chromium the script https://developer.chrome.com/extensions/crx
  has been integrated

Eventually the user can change the name of the generated extensions
files with the --extensions file.
On the chrome_extension/ directory file named "${EXTENSION_NAME}.pub"
and "${EXTENSION_NAME}.sig" will be removed when the script terminated
(or get killed).